Move the unit tests to separate file in "signature-bridge"
