+++ contracts/signature-bridge/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod contract;
 pub mod state;
+
+#[cfg(test)]
+pub mod tests;
